+++ e2e/e2etests/test_eth_withdraw.go
@@ -27,7 +27,7 @@ func TestETHWithdraw(r *runner.E2ERunner, args []string) {
 	// wait for the cctx to be mined
 	cctx := utils.WaitCctxMinedByInboundHash(r.Ctx, tx.Hash().Hex(), r.CctxClient, r.Logger, r.CctxTimeout)
 	r.Logger.CCTX(*cctx, "withdraw")
-	require.Equal(r, crosschaintypes.CctxStatus_OutboundMined, cctx.CctxStatus.Status)
+	utils.RequireCCTXStatus(r, cctx, crosschaintypes.CctxStatus_OutboundMined)
 
 	// check the balance was updated, we just check newBalance is greater than oldBalance because of the gas fee
 	newBalance, err := r.EVMClient.BalanceAt(r.Ctx, r.EVMAddress(), nil)