+++ e2e/e2etests/test_sui_token_withdraw_and_call.go
@@ -32,6 +32,8 @@ func TestSuiTokenWithdrawAndCall(r *runner.E2ERunner, args []string) {
 	// create the on_call payload
 	authorizedSender := r.EVMAddress()
 	payloadOnCall := r.SuiCreateExampleWACPayload(authorizedSender, suiAddress)
+	message, err := payloadOnCall.PackABI()
+	require.NoError(r, err)
 
 	// ACT
 	// approve both SUI gas budget token and fungible token ZRC20
@@ -43,7 +45,7 @@ func TestSuiTokenWithdrawAndCall(r *runner.E2ERunner, args []string) {
 		targetPackageID,
 		amount,
 		r.SuiTokenZRC20Addr,
-		payloadOnCall,
+		message,
 		gasLimit,
 		gatewayzevm.RevertOptions{OnRevertGasLimit: big.NewInt(0)},
 	)
@@ -53,7 +55,7 @@ func TestSuiTokenWithdrawAndCall(r *runner.E2ERunner, args []string) {
 	// wait for the cctx to be mined
 	cctx := utils.WaitCctxMinedByInboundHash(r.Ctx, tx.Hash().Hex(), r.CctxClient, r.Logger, r.CctxTimeout)
 	r.Logger.CCTX(*cctx, "withdraw_and_call")
-	require.EqualValues(r, crosschaintypes.CctxStatus_OutboundMined, cctx.CctxStatus.Status)
+	utils.RequireCCTXStatus(r, cctx, crosschaintypes.CctxStatus_OutboundMined)
 
 	// check the balance after the withdraw
 	balanceAfter := r.SuiGetFungibleTokenBalance(signer.Address())