+++ e2e/e2etests/test_zevm_to_evm_call.go
@@ -13,7 +13,8 @@ import (
 )
 
 func TestZEVMToEVMCall(r *runner.E2ERunner, args []string) {
-	require.Len(r, args, 0)
+	require.Len(r, args, 1)
+	gasLimit := utils.ParseBigInt(r, args[0])
 
 	payload := randomPayload(r)
 
@@ -29,12 +30,13 @@ func TestZEVMToEVMCall(r *runner.E2ERunner, args []string) {
 		gatewayzevm.RevertOptions{
 			OnRevertGasLimit: big.NewInt(0),
 		},
+		gasLimit,
 	)
 
 	// wait for the cctx to be mined
 	cctx := utils.WaitCctxMinedByInboundHash(r.Ctx, tx.Hash().Hex(), r.CctxClient, r.Logger, r.CctxTimeout)
 	r.Logger.CCTX(*cctx, "call")
-	require.Equal(r, crosschaintypes.CctxStatus_OutboundMined, cctx.CctxStatus.Status)
+	utils.RequireCCTXStatus(r, cctx, crosschaintypes.CctxStatus_OutboundMined)
 
 	// check the payload was received on the contract
 	r.AssertTestDAppEVMCalled(true, payload, big.NewInt(0))