+++ e2e/e2etests/test_zevm_to_evm_call_revert.go
@@ -14,7 +14,8 @@ import (
 )
 
 func TestZEVMToEVMCallRevert(r *runner.E2ERunner, args []string) {
-	require.Len(r, args, 0)
+	require.Len(r, args, 1)
+	gasLimit := utils.ParseBigInt(r, args[0])
 
 	payload := randomPayload(r)
 
@@ -30,12 +31,13 @@ func TestZEVMToEVMCallRevert(r *runner.E2ERunner, args []string) {
 			RevertMessage:    []byte(payload),
 			OnRevertGasLimit: big.NewInt(200000),
 		},
+		gasLimit,
 	)
 
 	// wait for the cctx to be mined
 	cctx := utils.WaitCctxMinedByInboundHash(r.Ctx, tx.Hash().Hex(), r.CctxClient, r.Logger, r.CctxTimeout)
 	r.Logger.CCTX(*cctx, "call")
-	require.Equal(r, crosschaintypes.CctxStatus_Reverted, cctx.CctxStatus.Status)
+	utils.RequireCCTXStatus(r, cctx, crosschaintypes.CctxStatus_Reverted)
 
 	r.AssertTestDAppZEVMCalled(true, payload, big.NewInt(0))