+++ e2e/e2etests/test_zevm_to_evm_call_revert_and_abort.go
@@ -15,7 +15,8 @@ import (
 )
 
 func TestZEVMToEVMCallRevertAndAbort(r *runner.E2ERunner, args []string) {
-	require.Len(r, args, 0)
+	require.Len(r, args, 1)
+	gasLimit := utils.ParseBigInt(r, args[0])
 
 	r.ApproveETHZRC20(r.GatewayZEVMAddr)
 
@@ -34,12 +35,13 @@ func TestZEVMToEVMCallRevertAndAbort(r *runner.E2ERunner, args []string) {
 			OnRevertGasLimit: big.NewInt(200000),
 			AbortAddress:     testAbortAddr,
 		},
+		gasLimit,
 	)
 
 	// wait for the cctx to be mined
 	cctx := utils.WaitCctxMinedByInboundHash(r.Ctx, tx.Hash().Hex(), r.CctxClient, r.Logger, r.CctxTimeout)
 	r.Logger.CCTX(*cctx, "call")
-	require.Equal(r, crosschaintypes.CctxStatus_Aborted, cctx.CctxStatus.Status)
+	utils.RequireCCTXStatus(r, cctx, crosschaintypes.CctxStatus_Aborted)
 
 	// check onAbort was called
 	aborted, err := testAbort.IsAborted(&bind.CallOpts{})