+++ x/fungible/types/codec.go
@@ -20,6 +20,7 @@ func RegisterCodec(cdc *codec.LegacyAmino) {
 	cdc.RegisterConcrete(&MsgUpdateGatewayContract{}, "fungible/UpdateGatewayContract", nil)
 	cdc.RegisterConcrete(&MsgUpdateZRC20Name{}, "fungible/UpdateZRC20Name", nil)
 	cdc.RegisterConcrete(&MsgBurnFungibleModuleAsset{}, "fungible/BurnFungibleModuleAsset", nil)
+	cdc.RegisterConcrete(&MsgUpdateGatewayGasLimit{}, "fungible/UpdateGatewayGasLimit", nil)
 }
 
 func RegisterInterfaces(registry cdctypes.InterfaceRegistry) {
@@ -36,6 +37,7 @@ func RegisterInterfaces(registry cdctypes.InterfaceRegistry) {
 		&MsgUpdateGatewayContract{},
 		&MsgUpdateZRC20Name{},
 		&MsgBurnFungibleModuleAsset{},
+		&MsgUpdateGatewayGasLimit{},
 	)
 
 	msgservice.RegisterMsgServiceDesc(registry, &_Msg_serviceDesc)