+++ changelog.md
@@ -5,6 +5,7 @@
 ### Features
 
 * [4064](https://github.com/zeta-chain/node/pull/4064) - add support for withdraws using the new v2 connector contract
+* [4153](https://github.com/zeta-chain/node/pull/4153) - make the gas limit used for gateway calls a configurable parameter
 
 ### Refactor
 
@@ -22,6 +23,7 @@
 
 * [4071](https://github.com/zeta-chain/node/pull/4071) - use v2 connector contract in admin e2e tests
 * [4113](https://github.com/zeta-chain/node/pull/4113) - fix Solana flaky depositAndCall e2e tests in live networks
+* [4142](https://github.com/zeta-chain/node/pull/4142) - fix Solana flaky SPL deposit e2e test in live networks
 
 ## v33.0.0
 
@@ -35,6 +37,7 @@
 * [3991](https://github.com/zeta-chain/node/pull/3991) - migrate from Ethermint to new Cosmos EVM module
 * [4077](https://github.com/zeta-chain/node/pull/4077) - add Cosmos EVM default precompiles support
 * [4068](https://github.com/zeta-chain/node/pull/4068) - add support for hostname in addition to public IP for zetaclient to connect to zetacore
+* [4087](https://github.com/zeta-chain/node/pull/4087) - re-enable Sui authenticated call and adopt MessageContext ID as a gateway dynamic field
 
 ### Fixes